fix - fixed some issues

+++ config/jwt.php
@@ -5,14 +5,14 @@ return [
     'secret' => env('JWT_SECRET'),
 
     'keys' => [
-        'public'  => env('JWT_PUBLIC_KEY'),
+        'public' => env('JWT_PUBLIC_KEY'),
         'private' => env('JWT_PRIVATE_KEY'),
         'passphrase' => env('JWT_PASSPHRASE'),
     ],
 
-    'ttl' => env('JWT_TTL', 1440),
+    'ttl' => (int) env('JWT_TTL', 1440),
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160),
 
     'algo' => env('JWT_ALGO', 'HS256'),
